requirements.txt
pytest
selenium
pony
suds-jurko
jsonpickle
